+++ test/Models/TerriaSpec.ts
@@ -25,6 +25,7 @@ import Feature from "../../lib/Models/Feature";
 import {
   isInitFromData,
   isInitFromDataPromise,
+  isInitFromOptions,
   isInitFromUrl
 } from "../../lib/Models/InitSource";
 import Terria from "../../lib/Models/Terria";
@@ -141,6 +142,47 @@ describe("Terria", function() {
       jasmine.Ajax.uninstall();
     });
 
+    it("works with initializationUrls and initFragmentPaths", async function() {
+      expect(terria.initSources.length).toEqual(0);
+
+      jasmine.Ajax.stubRequest("path/to/config/configUrl.json").andReturn({
+        responseText: JSON.stringify({
+          initializationUrls: ["something"],
+          parameters: {
+            applicationUrl: "https://application.url/",
+            initFragmentPaths: [
+              "path/to/init/",
+              "https://hostname.com/some/other/path/"
+            ]
+          }
+        })
+      });
+
+      await terria.start({
+        configUrl: `path/to/config/configUrl.json`,
+        i18nOptions
+      });
+
+      expect(terria.initSources.length).toEqual(1);
+
+      const initSource = terria.initSources[0];
+      expect(isInitFromOptions(initSource)).toBeTruthy();
+
+      if (!isInitFromOptions(initSource))
+        throw "Init source is not from options";
+
+      // Note: initFragmentPaths in `initializationUrls` are resolved to the base URL of configURL
+      // - which is path/to/config/
+      expect(
+        initSource.options.map(source =>
+          isInitFromUrl(source) ? source.initUrl : ""
+        )
+      ).toEqual([
+        "path/to/config/path/to/init/something.json",
+        "https://hostname.com/some/other/path/something.json"
+      ]);
+    });
+
     describe("via loadMagdaConfig", function() {
       it("should dereference uniqueId to `/`", function(done) {
         expect(terria.catalog.group.uniqueId).toEqual("/");
@@ -165,6 +207,7 @@ describe("Terria", function() {
             done.fail(error);
           });
       });
+
       it("works with basic initializationUrls", function(done) {
         jasmine.Ajax.stubRequest(/.*api\/v0\/registry.*/).andReturn({
           // terria's "Magda derived url"
@@ -195,6 +238,7 @@ describe("Terria", function() {
             done.fail(error);
           });
       });
+
       it("works with v7initializationUrls", async function() {
         jasmine.Ajax.stubRequest(/.*api\/v0\/registry.*/).andReturn({
           // terria's "Magda derived url"
@@ -332,6 +376,59 @@ describe("Terria", function() {
   });
 
   describe("updateApplicationUrl", function() {
+    it("works with initializationUrls and initFragmentPaths", async function() {
+      expect(terria.initSources.length).toEqual(0);
+
+      jasmine.Ajax.install();
+      // Fail all requests by default.
+      jasmine.Ajax.stubRequest(/.*/).andError({});
+
+      jasmine.Ajax.stubRequest("path/to/config/configUrl.json").andReturn({
+        responseText: JSON.stringify({
+          initializationUrls: ["something"],
+          parameters: {
+            applicationUrl: "https://application.url/",
+            initFragmentPaths: [
+              "path/to/init/",
+              "https://hostname.com/some/other/path/"
+            ]
+          }
+        })
+      });
+
+      await terria.start({
+        configUrl: `path/to/config/configUrl.json`,
+        i18nOptions
+      });
+
+      await terria.updateApplicationUrl(
+        "https://application.url/#someInitHash"
+      );
+
+      console.log(terria.initSources);
+
+      expect(terria.initSources.length).toEqual(2);
+
+      const initSource = terria.initSources[1];
+      expect(isInitFromOptions(initSource)).toBeTruthy();
+
+      if (!isInitFromOptions(initSource))
+        throw "Init source is not from options";
+
+      // Note: initFragmentPaths in hash parameters are resolved to the base URL of application URL
+      // - which is https://application.url/
+      expect(
+        initSource.options.map(source =>
+          isInitFromUrl(source) ? source.initUrl : ""
+        )
+      ).toEqual([
+        "https://application.url/path/to/init/someInitHash.json",
+        "https://hostname.com/some/other/path/someInitHash.json"
+      ]);
+
+      jasmine.Ajax.uninstall();
+    });
+
     describe("test via serialise & load round-trip", function() {
       let newTerria: Terria;
       let viewState: ViewState;
@@ -501,6 +598,7 @@ describe("Terria", function() {
           storyRouteUrlPrefix: "test/stories/TerriaJS%20App/"
         });
       });
+
       it("sets playStory to 1", async function() {
         await terria.updateApplicationUrl(
           new URL("story/my-story", document.baseURI).toString()